# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"joao-magalhaes","description":"joao-magalhaes","contact":{"name":"joao-magalhaes","email":"joao.jmagalhaes@mercadolivre.com"},"version":"last"},"servers":[{"url":"https://null_joao-magalhaes.furyapps.io","description":"Scope null"}],"paths":{"/api/v1/fresh-products/orders/":{"get":{"tags":["buy-order-controller"],"operationId":"getOrderItems","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BuyOrderProductsView"}}}}}}},"put":{"tags":["buy-order-controller"],"operationId":"updateOrderItems","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyOrderForm"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"201":{"description":"Created"}}},"post":{"tags":["buy-order-controller"],"operationId":"createBuyOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyOrderForm"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BuyOrderView"}}}}}}},"/api/v1/fresh-products/inboundorder/":{"put":{"tags":["insert-batch-controller"],"operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutInboundForm"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockView"}}}}}}},"post":{"tags":["insert-batch-controller"],"operationId":"insert","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderForm"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockView"}}}}}}}},"/auth":{"post":{"tags":["authenticate-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginForm"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenView"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/warehouseStatistics/{id}":{"get":{"tags":["warehouse-controller"],"operationId":"findWarehouse","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WarehouseStatisticsView"}}}}}}},"/api/v1/fresh-products/{category}":{"get":{"tags":["purchase-controller"],"operationId":"listAllByCategory","parameters":[{"name":"category","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductView"}}}}}}}},"/api/v1/fresh-products/price/{order}":{"get":{"tags":["purchase-controller"],"operationId":"listAllAndSortByPrice","parameters":[{"name":"order","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductView"}}}}}}}},"/api/v1/fresh-products/name/{name}":{"get":{"tags":["find-product-by-name-controller"],"operationId":"listAllByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductView"}}}}}}},"/api/v1/fresh-products/complete/{id}":{"get":{"tags":["complete-buy-order-controller"],"operationId":"completeBuyOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BuyOrderCompletedView"}}}}}}},"/api/v1/fresh-products/":{"get":{"tags":["purchase-controller"],"operationId":"listAll","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductView"}}}}}}}},"/api/v1/fresh-products/due-date/overdue":{"delete":{"tags":["due-date-controller"],"operationId":"deleteOverDueStocks","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OverDueView"}}}}}}},"/api/v1/fresh-products/warehouse/{querytype}":{"get":{"tags":["product-location-controller"],"operationId":"listByWarehouse_5","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWithIdWarehouseView"}}}}}},"put":{"tags":["product-location-controller"],"operationId":"listByWarehouse_6","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWithIdWarehouseView"}}}}}},"post":{"tags":["product-location-controller"],"operationId":"listByWarehouse","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWithIdWarehouseView"}}}}}},"delete":{"tags":["product-location-controller"],"operationId":"listByWarehouse_4","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWithIdWarehouseView"}}}}}},"options":{"tags":["product-location-controller"],"operationId":"listByWarehouse_3","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWithIdWarehouseView"}}}}}},"head":{"tags":["product-location-controller"],"operationId":"listByWarehouse_1","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWithIdWarehouseView"}}}}}},"patch":{"tags":["product-location-controller"],"operationId":"listByWarehouse_2","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWithIdWarehouseView"}}}}}}},"/api/v1/fresh-products/list":{"get":{"tags":["product-location-controller"],"operationId":"listById_5","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductLocationView"}}}}}}},"put":{"tags":["product-location-controller"],"operationId":"listById_6","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductLocationView"}}}}}}},"post":{"tags":["product-location-controller"],"operationId":"listById","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductLocationView"}}}}}}},"delete":{"tags":["product-location-controller"],"operationId":"listById_4","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductLocationView"}}}}}}},"options":{"tags":["product-location-controller"],"operationId":"listById_3","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductLocationView"}}}}}}},"head":{"tags":["product-location-controller"],"operationId":"listById_1","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductLocationView"}}}}}}},"patch":{"tags":["product-location-controller"],"operationId":"listById_2","parameters":[{"name":"querytype","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductLocationView"}}}}}}}},"/api/v1/fresh-products/due-date/{querytype}":{"get":{"tags":["due-date-controller"],"operationId":"listByDueDate_5","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"put":{"tags":["due-date-controller"],"operationId":"listByDueDate_6","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"post":{"tags":["due-date-controller"],"operationId":"listByDueDate","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"delete":{"tags":["due-date-controller"],"operationId":"listByDueDate_4","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"options":{"tags":["due-date-controller"],"operationId":"listByDueDate_3","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"head":{"tags":["due-date-controller"],"operationId":"listByDueDate_1","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"patch":{"tags":["due-date-controller"],"operationId":"listByDueDate_2","parameters":[{"name":"querytype","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}}},"/api/v1/fresh-products/due-date/list":{"get":{"tags":["due-date-controller"],"operationId":"listByDueDate_12","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"put":{"tags":["due-date-controller"],"operationId":"listByDueDate_13","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"post":{"tags":["due-date-controller"],"operationId":"listByDueDate_7","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"delete":{"tags":["due-date-controller"],"operationId":"listByDueDate_11","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"options":{"tags":["due-date-controller"],"operationId":"listByDueDate_10","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"head":{"tags":["due-date-controller"],"operationId":"listByDueDate_8","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}},"patch":{"tags":["due-date-controller"],"operationId":"listByDueDate_9","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiException"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DueDateView"}}}}}}}}},"components":{"schemas":{"ApiException":{"type":"object","properties":{"cause":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"code":{"type":"string"},"description":{"type":"string"},"statusCode":{"type":"integer","format":"int32"},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"BuyOrderForm":{"required":["buyerId","products"],"type":"object","properties":{"buyerId":{"type":"integer","format":"int64"},"products":{"type":"array","items":{"$ref":"#/components/schemas/BuyProductsForm"}}}},"BuyProductsForm":{"required":["productId","quantity"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"}}},"PutInboundForm":{"required":["batch_stock","orderNumber","section"],"type":"object","properties":{"section":{"$ref":"#/components/schemas/SectionForm"},"orderNumber":{"type":"integer","format":"int64"},"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/PutStockForm"}}}},"PutStockForm":{"required":["currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","minimumTemperature","number","productId"],"type":"object","properties":{"number":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"currentTemperature":{"type":"number","format":"float"},"minimumTemperature":{"type":"number","format":"float"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string"},"manufacturingTime":{"type":"string"},"dueDate":{"type":"string"}}},"SectionForm":{"required":["sectionCode","warehouseCode"],"type":"object","properties":{"sectionCode":{"type":"string"},"warehouseCode":{"type":"string"}}},"StockView":{"type":"object","properties":{"number":{"type":"integer","format":"int64"},"productId":{"type":"string"},"currentTemperature":{"type":"number","format":"float"},"minimumTemperature":{"type":"number","format":"float"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string"},"manufacturingTime":{"type":"string"},"dueDate":{"type":"string"}}},"LoginForm":{"required":["password","user_name"],"type":"object","properties":{"password":{"type":"string"},"user_name":{"type":"string"}}},"TokenView":{"type":"object","properties":{"token":{"type":"string"},"type":{"type":"string"}}},"BuyOrderView":{"type":"object","properties":{"totalPrice":{"type":"object","properties":{"opaque":{"type":"number","format":"double"},"acquire":{"type":"number","format":"double"},"release":{"type":"number","format":"double","writeOnly":true},"plain":{"type":"number","format":"double"}}},"returnView":{"type":"object","additionalProperties":{"type":"string"}}}},"InboundOrderForm":{"required":["batch_stock","section"],"type":"object","properties":{"section":{"$ref":"#/components/schemas/SectionForm"},"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/StockForm"}},"order_date":{"type":"string"}}},"StockForm":{"required":["currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","minimumTemperature","productId"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"currentTemperature":{"type":"number","format":"float"},"minimumTemperature":{"type":"number","format":"float"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string"},"manufacturingTime":{"type":"string"},"dueDate":{"type":"string"}}},"WarehouseStatisticsView":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"qtdSection":{"type":"integer","format":"int32"},"totalSize":{"type":"number","format":"double"},"totalProducts":{"type":"integer","format":"int32"}}},"ProductView":{"type":"object","properties":{"name":{"type":"string"},"price":{"type":"number","format":"double"},"category":{"type":"string"}}},"BuyOrderProductsView":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"}}},"BuyOrderCompletedView":{"type":"object","properties":{"buyOrderId":{"type":"string"},"orderStatus":{"type":"string"},"orderDate":{"type":"string"},"cartItemList":{"type":"array","items":{"$ref":"#/components/schemas/CartItem"}},"valid":{"type":"string"}}},"CartItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"}}},"OverDueView":{"type":"object","properties":{"deletedQuantity":{"type":"integer","format":"int64"},"deleted_quantity":{"type":"integer","format":"int64"}}},"ProductInWarehouseView":{"type":"object","properties":{"warehouseCode":{"type":"string"},"totalQuantity":{"type":"integer","format":"int32"}}},"ProductWithIdWarehouseView":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"productInWarehouseViewList":{"type":"array","items":{"$ref":"#/components/schemas/ProductInWarehouseView"}}}},"ProductLocationView":{"type":"object","properties":{"sectionCode":{"type":"string"},"warehouseCode":{"type":"string"},"productId":{"type":"integer","format":"int64"},"batchNumber":{"type":"integer","format":"int64"},"currentQuantity":{"type":"integer","format":"int32"},"dueDate":{"type":"string"}}},"DueDateView":{"type":"object","properties":{"batchNumber":{"type":"integer","format":"int32"},"productId":{"type":"string"},"category":{"type":"string"},"dueDate":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"section":{"type":"string"},"warehouse":{"type":"string"}}}}}}